+++ app/controllers/pdl.py
@@ -1022,8 +1022,10 @@ def generate_title_iv_state_distribution_response(program_id, start_year, end_ye
         Program.name.label('programName'),
         Payment.year.label('year'),
         Payment.payment.label('totalPaymentInDollars'),
-        Payment.recipient_count.label('totalCounts')).join(
-        Program, Payment.program_id == Program.id).filter(
+        Payment.recipient_count.label('totalCounts')
+    ).join(
+        Program, Payment.program_id == Program.id
+    ).filter(
         Payment.program_id == program_id,
         Payment.year.between(start_year, end_year)
     )
@@ -1035,7 +1037,7 @@ def generate_title_iv_state_distribution_response(program_id, start_year, end_ye
     program_response_dict = defaultdict(list)
     year_dict = defaultdict(list)
     total_national_payment_dict = defaultdict(float)
-    total_national_count_dict = defaultdict(int)
+    total_national_average_monthly_participation_dict = defaultdict(int)
 
     # aggregate data
     for record in program_result:
@@ -1052,20 +1054,18 @@ def generate_title_iv_state_distribution_response(program_id, start_year, end_ye
 
         # update total national payment and count for the year
         total_national_payment_dict[year] += payment
-        total_national_count_dict[year] += recipient_count
+        total_national_average_monthly_participation_dict[year] += recipient_count
 
     # create output response dictionary
     for year, data in year_dict.items():
         program_response_dict[year] = []
-        total_payment_nationwide = total_national_payment_dict[year]
-        total_participation_nationwide = total_national_count_dict[year]
         for item in data:
             state = item['state']
             total_payment = item['totalPaymentInDollars']
             average_participation = item['averageMonthlyParticipation']
 
-            total_payment_percentage = (total_payment / total_payment_nationwide) * 100
-            average_participation_percentage = (average_participation / total_participation_nationwide) * 100
+            total_payment_percentage = (total_payment / total_national_payment_dict[year]) * 100
+            average_participation_percentage = (average_participation / total_national_average_monthly_participation_dict[year]) * 100
 
             program_response_dict[year].append({
                 'state': state,
@@ -1093,25 +1093,25 @@ def generate_title_iv_state_distribution_response(program_id, start_year, end_ye
 
     # calculate the sum of total national payment and count for all years
     total_national_payment = sum(total_national_payment_dict.values())
-    total_national_count = sum(total_national_count_dict.values())
+    national_average_monthly_participation_across_years = (sum(total_national_average_monthly_participation_dict.values()))/(end_year - start_year + 1)
 
     # add the total data for each state to the program_response_dict
     total_states_data = []
     for state, total_data in state_total_data.items():
         total_payment = total_data['totalPaymentInDollars']
-        total_participation = total_data['averageMonthlyParticipation']
+        average_monthly_participation = total_data['averageMonthlyParticipation']/(end_year - start_year + 1)
 
         #  calculate percentage using the total national payment and count across all years
         total_payment_percentage = (total_payment / total_national_payment) * 100
-        total_participation_percentage = (total_participation / total_national_count) * 100
+        average_monthly_participation_percentage = (average_monthly_participation / national_average_monthly_participation_across_years) * 100
 
         # append total data for each state to the list
         total_states_data.append({
             'state': state,
             'totalPaymentInDollars': total_payment,
             'totalPaymentInPercentageNationwide': round(total_payment_percentage, 2),
-            'averageMonthlyParticipation': total_participation,
-            'averageMonthlyParticipationInPercentageNationwide': round(total_participation_percentage, 2)
+            'averageMonthlyParticipation': round(average_monthly_participation),
+            'averageMonthlyParticipationInPercentageNationwide': round(average_monthly_participation_percentage, 2)
         })
 
     # Sort the total states data by decreasing order of total payment in dollars